+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Support for iPad multitasking / multiple windows. State restoration covers the open document, selected hashtag, and opened page. Maybe someday I'll add more UI state.
+- Add hashtag typeahead completion
 
 ## [0.38.0] - 2020-01-05